Added an option to "opt-out" of deleting the project. As a failsafe if you typed something you have been working for a while.

+++ Source/gen.cs
@@ -32,9 +32,15 @@ namespace generator
 					Generate(dir, input, config.assets);
 				else if (key == 'd' || key == 'D')
 				{
-					//Console.Write("Are you sure you want to ");
-					DirectoryInfo dirInfo = new DirectoryInfo(dir);
-					dirInfo.Delete(true);
+					Console.Write("Are you sure you want to delete '{0}' project? (y/n)", dir);
+					key = (char)Console.Read();
+					if (key == 'y' || key == 'Y')
+					{
+						DirectoryInfo dirInfo = new DirectoryInfo(dir);
+						dirInfo.Delete(true);
+					}
+					else if (key == 'n' || key == 'N')
+						return;
 				}
 				else
 					Console.WriteLine("This is not an option!");
